Modified README.md, added rti_scenarios json
Added beginning content to `README.md`

Added `deprecated_rti_scenarios.json`
This file is from .../fh_models/who_tool/templates/RTI-Kenya/RTI-C50-Scenarios.json
It was dated 2022-02-25 and i'm assuming it's the most recent set of
scenarios that we used for the Kenya work with RTI and Brian.

I will use this file to build the botech model file, and the run it,
comparing the new results with old results, which I will need to find.